chitchat_server/requirements.txt
Flask-SQLAlchemy
Flask
psycopg2-binary
gunicorn
